+++ CHANGELOG.md
@@ -1,11 +1,11 @@
-# Changelog for FormatChange
+# Changelog
 
 ## Unreleased...
 * ... <!-- Add new lines here. Version number will be decided later -->
 
 
 ## 1.7.0
-_2022-12-21_
+_2022-02-21_
 * feat: Add TypeScript type declarations
 
 ## 1.6.3 / 1.6.4