+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased...
 
 - ... <!-- Add new lines here. Version number will be decided later -->
+- fix: Bugs in Type declarations
 
 ## 1.7.0